requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiokafka==0.12.0
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
multidict==6.4.4
packaging==25.0
propcache==0.3.2
PyYAML==6.0.2
typing_extensions==4.14.0
yarl==1.20.1
